requirements.txt
pyTelegramBotAPI==4.14.0
Flask==2.3.3
python-dotenv==1.0.0
psutil==5.9.5
Werkzeug==2.3.7
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.3.0
idna==3.4
requests==2.31.0
urllib3==2.0.6
gunicorn==21.2.0